pyproject.toml
[tool.poetry]
name = "alegrosz"
version = "0.1.0"
description = ""
authors = ["Bartosz Baran <bartosz.baran192@gmail.com>"]
license = "Beerware"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.1.7"
djangorestframework = "3.14.0"
pytest = "7.2.2"
pytest-django = "4.5.2"
mypy = "1.1.1"
black = "23.1.0"
flake8 = "6.0.0"
Sphinx = "6.1.3"
django-stubs = "1.15.0"
djangorestframework-stubs = "1.9.1"
pre-commit = "3.1.1"
renku-sphinx-theme = "0.2.1"
sphinx-rtd-theme = "1.2.0"
pillow = "9.4.0"
pytest-faker = "2.0.0"
faker-commerce = "1.0.3"
pytest-cov = "4.0.0"
coverage = "7.2.1"
pytest-factoryboy = "2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "test_*.py"

addopts = [
    "--cov=.",
    "--cov-config=.coveragerc",
    "--cov-report",
    "term-missing:skip-covered",
    "--cov-fail-under=80"
]

[tool.pytest.markers]
slow = "Run slow unit test e.g. for models"
unit = "Run unit tests"
smoke = "Run core unit tests"
dbtest = "Run database tests"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
ignore = ['E231', 'E241', 'F401']
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 120
count = true
